product_service/api/pyproject.toml
[tool.poetry]
name = "product-service"
version = "0.1.0"
description = "product microservice"
authors = ["blazej <bbielski@lyck.pl>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
Django = "5.1.2"
djangorestframework = "3.15.2"
psycopg = "3.2.3"
django-debug-toolbar = "4.4.6"
drf-spectacular = "0.27.2"
django-filter = "24.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
